🔨 ensure title and subtitle in grapher config respect fallbacks
Also improved the refreshExplorerViews script so it can run only a subset of explorers

+++ packages/@ourworldindata/grapher/src/core/Grapher.tsx
@@ -532,7 +532,7 @@ export class GrapherState {
         }
     }
 
-    toObject(): GrapherInterface {
+    toObject(deleteUnchanged: boolean = true): GrapherInterface {
         const obj: GrapherInterface = objectWithPersistablesToObject(
             this,
             grapherKeysToSerialize
@@ -541,7 +541,9 @@ export class GrapherState {
         obj.selectedEntityNames = this.selection.selectedEntityNames
         obj.focusedSeriesNames = this.focusArray.seriesNames
 
-        deleteRuntimeAndUnchangedProps(obj, defaultObject)
+        if (deleteUnchanged) {
+            deleteRuntimeAndUnchangedProps(obj, defaultObject)
+        }
 
         // always include the schema, even if it's the default
         obj.$schema = this.$schema || latestGrapherConfigSchema
@@ -2181,7 +2183,7 @@ export class GrapherState {
             ? this.yColumnsFromDimensions
             : this.table.getColumns(autoDetectYColumnSlugs(this))
     }
-    @computed private get defaultTitle(): string {
+    @computed get defaultTitle(): string {
         const yColumns = this.yColumnsFromDimensionsOrSlugsOrAuto
 
         if (this.isScatter)
